Review fixes: lock-file theft, three collection regressions, shared statements, and the docs
An adversarial review of this branch against master -- nine reviewers over
the full diff, every finding verified by two independent agents trying to
refute it -- confirmed 22 findings. This commit fixes them.

CRITICAL -- closing a lock-failed database deleted the RUNNING instance's
lock files. _lock_single_database() stored the lock path before the
O_EXCL create, so on a failed acquisition the field pointed at the OTHER
process's live lock file -- and _database_free() unlinks those fields
unconditionally. Unreachable at the merge base (the old retry loop leaked
the failed db), made reachable by this branch's own leak fix, which
closes it before retrying: a second instance would delete the first's
data lock, retry, delete its library lock, retry, and open the same
database -- silently, in headless mode. The path now survives only when
this process created the file, which is also what makes the GUI Retry
button's documented "without touching any lock file" true.

Also in database.c: on a corrupt-database retry, `library` could still
alias the freed previous migrated name (use-after-free straight into the
migration probe); the previous allocation now lives until a new name
actually replaces it, and three early returns that leaked it free it.

Three collection regressions against the original:

  * An OR rule's value list was restricted by every rule INCLUDING
    itself; the original applied no rule at all ("don't limit the
    collection for OR"). _extended_where_excluding(apply_exclude=FALSE)
    now composes no restriction.
  * The generation counter bumped on every recomposition, so the
    thumbtable's collection hash fired -- a full grid reset -- on every
    RELOAD (rating, tag, import batch), where hashing the query text was
    a no-op by that enum's own definition. It now advances only when the
    composed text changes, which is what its consumers were promised.
  * dt_collection_get_count() in culling mode reported the FULL
    collection: the refresh counted before the caller's culling
    restriction deleted the non-selected rows.
    restrict_to_selection() recounts after its DELETE.

Cross-thread shared statements: six repositories kept 22 cached
sqlite3_stmt handles with no mutex -- carried over from the pre-PR files,
but this module's own README claimed lock coverage, and its three
mutexed repositories exist because the same sharing is a race: two
threads stepping one statement interleave both row cursors, and the lazy
first-use prepare can prepare twice. All 22 (metadata, colorlabel,
selection, tag, preset, collection_query) now prepare and finalise per
call -- removing the shared state instead of guarding it, the idiom most
of this module already uses. Their cleanups stay, empty, because
dt_database_close() runs every repository's cleanup.

GPX map markers: the geotagging panel's image list moved from an SQL scan
(ascending imgid, the table's INTEGER PRIMARY KEY) to the in-memory
selection (pick order), but marker grouping collapses same-location
images by comparing each entry with its list NEIGHBOUR. The list is
sorted ascending again -- the same order dependence the group-leader fix
in gui/actions/image.c already handled for its own consumer.

Behaviour restored in two more places: focal-length auto-apply bounds are
REAL columns and round-tripped as doubles through the edit dialog (the
conditions struct read them as int, truncating the stored bound on every
open-and-save); dt_map_location_get_images_count() answers 0 again on a
failed read (the shared repository function keeps dt_tag_remove()'s -1).

And the documentation now tells the truth where it did not: store()'s
route to the colour labels, copy_to_film's THREE satellite tables
(module_order travels with the history, on purpose), the geo-bounds
±INFINITY seeding, write_mask_item's guard (fixed on this branch, not
"preserved"), the history preamble's mutex contract (foreach_row DOES
hold the mutex across callbacks -- documented as the constraint it is),
image_offset's 0-not--1, list_for_module's unordered plain branch, three
cleanup doc blocks reattached to the functions they describe, and the
README's _db_lock section no longer claims a per-query read side that
does not exist.

The review also REFUTED one finding and could not confirm three more --
recorded here so they are not re-found: the history foreach deadlock
claim (no current callback re-enters the repository), and the dropped
minors live in the review artifact.

ctest 17/17 on gcc and clang Debug; all four configurations and every
gate pass.

+++ src/database/preset_repository.h
@@ -126,10 +126,13 @@ void dt_module_preset_free(gpointer data);
 /**
  * @brief Every preset of `(operation, op_version)`.
  *
- * @param with_description also read the description column.
+ * @param with_description also read the description column. When FALSE the rows come back
+ *        UNORDERED and @p shipped_first is ignored -- faithful to the caller this branch
+ *        replaced (get_active_preset_name), which scans every row for a params match and
+ *        never cared about order.
  * @param shipped_first order the read-only presets before the user's, rather than after.
  *        A boolean rather than a sort direction: the caller used to interpolate "DESC" or
- *        "ASC" into the query text itself.
+ *        "ASC" into the query text itself. Only meaningful with @p with_description.
  */
 GList *dt_preset_repository_list_for_module(const char *operation, const int op_version,
                                             const gboolean with_description,
@@ -379,7 +382,10 @@ typedef struct dt_preset_conditions_t
   double iso_min, iso_max;
   double exposure_min, exposure_max;
   double aperture_min, aperture_max;
-  int focal_length_min, focal_length_max;
+  /* REAL columns, like the pairs above: read and bound as doubles so a fractional stored
+   * bound survives an open-and-save of the edit dialog. The dialog displays whole
+   * millimetres, but display rounding is the dialog's, not the store's. */
+  double focal_length_min, focal_length_max;
   int autoapply;
   int filter;
   int format;
